TOOL/updater/requirements.txt
esptool>=4.7
pyserial>=3.5
